requirements.txt
attrs==21.2.0
click==8.0.3
flasgger==0.9.5
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
mistune==0.8.4
psycopg2-binary==2.9.2
pyrsistent==0.18.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
